Fixing test failures. Given that these failures are reproduced only in CI build and that too after adding new tests that were setting processwide dpi awareness, I am guessing filaure is because of incorrect DPIawareness between the threads while running tests. Replacing process wide DPI awareness with thread.
